backend/requirements.txt
# Requirements mínimos - apenas dependências em uso
# Core Django
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1

# Base de dados
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Servidor de produção
gunicorn==21.2.0
whitenoise==6.6.0

# Configuração e HTTP requests
python-decouple==3.8
requests==2.31.0

# Serviços externos
twilio==8.10.0
Pillow==10.0.1

# Variáveis de ambiente
python-dotenv==1.0.0
python-decouple==3.8

# Utilitários básicos (usado em vários lugares)
python-dateutil==2.8.2

# Firebase e Google Cloud
# firebase-admin>=6.4.0
# google-cloud-storage>=2.14.0
# google-cloud-aiplatform>=1.38.0
# vertexai>=0.0.1
# google.generativeai>=0.3.0
